+++ src/mlpack/tests/hmm_test.cpp
@@ -1243,18 +1243,18 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMPredictTest)
     std::vector<DiagonalGMM> gmms(2);
     gmms[0] = DiagonalGMM(2, 2);
 
-    gmms[0].Component(0) = DiagCovGaussianDistribution("3.25 2.10",
+    gmms[0].Component(0) = DiagonalGaussianDistribution("3.25 2.10",
         "0.97 1.00");
-    gmms[0].Component(1) = DiagCovGaussianDistribution("5.03 7.28",
+    gmms[0].Component(1) = DiagonalGaussianDistribution("5.03 7.28",
         "1.20 0.89");
 
     gmms[1] = DiagonalGMM(3, 2);
     gmms[1].Weights() = arma::vec("0.3 0.2 0.5");
-    gmms[1].Component(0) = DiagCovGaussianDistribution("-2.48 -3.02",
+    gmms[1].Component(0) = DiagonalGaussianDistribution("-2.48 -3.02",
         "1.02 0.80");
-    gmms[1].Component(0) = DiagCovGaussianDistribution("-1.24 -2.40",
+    gmms[1].Component(0) = DiagonalGaussianDistribution("-1.24 -2.40",
         "0.85 0.78");
-    gmms[1].Component(0) = DiagCovGaussianDistribution("-5.68 -4.83",
+    gmms[1].Component(0) = DiagonalGaussianDistribution("-5.68 -4.83",
         "1.42 0.96");
 
     // Initial probabilities.
@@ -1314,14 +1314,14 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMPredictTest)
 BOOST_AUTO_TEST_CASE(DiagonalGMMHMMGenerateTest)
 {
   // Build the model.
-  HMM<DiagCovGaussianDistribution> hmm(3, DiagCovGaussianDistribution(2));
+  HMM<DiagonalGaussianDistribution> hmm(3, DiagonalGaussianDistribution(2));
   hmm.Transition() = arma::mat("0.2 0.3 0.8;"
                                "0.4 0.5 0.1;"
                                "0.4 0.2 0.1");
 
-  hmm.Emission()[0] = DiagCovGaussianDistribution("0.0 0.0", "1.0 0.7");
-  hmm.Emission()[1] = DiagCovGaussianDistribution("1.0 1.0", "0.7 0.5");
-  hmm.Emission()[2] = DiagCovGaussianDistribution("-3.0 2.0", "2.0 0.3");
+  hmm.Emission()[0] = DiagonalGaussianDistribution("0.0 0.0", "1.0 0.7");
+  hmm.Emission()[1] = DiagonalGaussianDistribution("1.0 1.0", "0.7 0.5");
+  hmm.Emission()[2] = DiagonalGaussianDistribution("-3.0 2.0", "2.0 0.3");
 
   // Now we will generate a long sequence.
   std::vector<arma::mat> observations(1);
@@ -1331,7 +1331,7 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMGenerateTest)
   hmm.Generate(10000, observations[0], states[0], 1);
 
   // Build the hmm2.
-  HMM<DiagCovGaussianDistribution> hmm2(3, DiagCovGaussianDistribution(2));
+  HMM<DiagonalGaussianDistribution> hmm2(3, DiagonalGaussianDistribution(2));
 
   // Now estimate the HMM from the generated sequence.
   hmm2.Train(observations, states);
@@ -1356,7 +1356,7 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMGenerateTest)
 BOOST_AUTO_TEST_CASE(DiagonalGMMHMMOneGaussianOneStateTrainingTest)
 {
   // Create a gaussian distribution with diagonal covariance.
-  DiagCovGaussianDistribution d("2.05 3.45", "0.89 1.05");
+  DiagonalGaussianDistribution d("2.05 3.45", "0.89 1.05");
 
   // Make a sequence of observations.
   std::vector<arma::mat> observations(1, arma::mat(2, 5000));
@@ -1394,10 +1394,10 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMOneGaussianUnlabeledTrainingTest)
 {
   // Create a sequence of DiagonalGMMs. Each GMM has one gaussian distribution.
   std::vector<DiagonalGMM> gmms(2, DiagonalGMM(1, 2));
-  gmms[0].Component(0) = DiagCovGaussianDistribution("1.25 2.10",
+  gmms[0].Component(0) = DiagonalGaussianDistribution("1.25 2.10",
       "0.97 1.00");
 
-  gmms[1].Component(0) = DiagCovGaussianDistribution("-2.48 -3.02",
+  gmms[1].Component(0) = DiagonalGaussianDistribution("-2.48 -3.02",
       "1.02 0.80");
 
   // Transition matrix.
@@ -1467,13 +1467,13 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMOneGaussianLabeledTrainingTest)
 {
   // Create a sequence of DiagonalGMMs.
   std::vector<DiagonalGMM> gmms(3, DiagonalGMM(1, 2));
-  gmms[0].Component(0) = DiagCovGaussianDistribution("5.25 7.10",
+  gmms[0].Component(0) = DiagonalGaussianDistribution("5.25 7.10",
       "0.97 1.00");
 
-  gmms[1].Component(0) = DiagCovGaussianDistribution("4.48 6.02",
+  gmms[1].Component(0) = DiagonalGaussianDistribution("4.48 6.02",
       "1.02 0.80");
 
-  gmms[2].Component(0) = DiagCovGaussianDistribution("-3.28 -5.30",
+  gmms[2].Component(0) = DiagonalGaussianDistribution("-3.28 -5.30",
       "0.87 1.05");
 
   // Transition matrix.
@@ -1550,15 +1550,15 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMMultipleGaussiansUnlabeledTrainingTest)
   // Create a sequence of DiagonalGMMs.
   std::vector<DiagonalGMM> gmms(2, DiagonalGMM(2, 2));
   gmms[0].Weights() = arma::vec("0.3 0.7");
-  gmms[0].Component(0) = DiagCovGaussianDistribution("8.25 7.10",
+  gmms[0].Component(0) = DiagonalGaussianDistribution("8.25 7.10",
       "0.97 1.00");
-  gmms[0].Component(1) = DiagCovGaussianDistribution("-3.03 -2.28",
+  gmms[0].Component(1) = DiagonalGaussianDistribution("-3.03 -2.28",
       "1.20 0.89");
 
   gmms[1].Weights() = arma::vec("0.4 0.6");
-  gmms[1].Component(0) = DiagCovGaussianDistribution("4.48 6.02",
+  gmms[1].Component(0) = DiagonalGaussianDistribution("4.48 6.02",
         "1.02 0.80");
-  gmms[1].Component(1) = DiagCovGaussianDistribution("-9.24 -8.40",
+  gmms[1].Component(1) = DiagonalGaussianDistribution("-9.24 -8.40",
         "0.85 1.58");
 
   // Transition matrix.
@@ -1657,15 +1657,15 @@ BOOST_AUTO_TEST_CASE(DiagonalGMMHMMMultipleGaussiansLabeledTrainingTest)
   // Create a sequence of DiagonalGMMs.
   std::vector<DiagonalGMM> gmms(2, DiagonalGMM(2, 2));
   gmms[0].Weights() = arma::vec("0.3 0.7");
-  gmms[0].Component(0) = DiagCovGaussianDistribution("2.25 5.30",
+  gmms[0].Component(0) = DiagonalGaussianDistribution("2.25 5.30",
       "0.97 1.00");
-  gmms[0].Component(1) = DiagCovGaussianDistribution("-3.15 -2.50",
+  gmms[0].Component(1) = DiagonalGaussianDistribution("-3.15 -2.50",
       "1.20 0.89");
 
   gmms[1].Weights() = arma::vec("0.4 0.6");
-  gmms[1].Component(0) = DiagCovGaussianDistribution("-4.48 -6.30",
+  gmms[1].Component(0) = DiagonalGaussianDistribution("-4.48 -6.30",
         "1.02 0.80");
-  gmms[1].Component(1) = DiagCovGaussianDistribution("5.24 2.40",
+  gmms[1].Component(1) = DiagonalGaussianDistribution("5.24 2.40",
         "0.85 1.58");
 
   // Transition matrix.